feat(model): 랜덤 번호 생성 기능 추가
