+++ src/main/java/lotto/model/Lotto.java
@@ -1,0 +1,33 @@
+package lotto.model;
+
+import camp.nextstep.edu.missionutils.Randoms;
+
+import java.util.List;
+
+public class Lotto {
+    private static final int LOTTO_START = 1;
+    private static final int LOTTO_END = 45;
+    private static final int LOTTO_SIZE = 6;
+
+    private final List<Integer> numbers;
+
+    public Lotto(List<Integer> numbers) {
+        validate(numbers);
+        this.numbers = numbers;
+    }
+
+    private void validate(List<Integer> numbers) {
+        if (numbers.size() != 6) {
+            throw new IllegalArgumentException();
+        }
+    }
+
+    public static Lotto create() {
+        List<Integer> lottoNumbers = pickUniqueRandomNumbers();
+        return new Lotto(lottoNumbers);
+    }
+
+    protected static List<Integer> pickUniqueRandomNumbers() {
+        return Randoms.pickUniqueNumbersInRange(LOTTO_START, LOTTO_END, LOTTO_SIZE);
+    }
+}
